backend/pyproject.toml
[project]
name = "dino-gallery-backend"
version = "0.1.0"
description = "AI-powered photo gallery backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "onnxruntime>=1.17.0",
    "onnx>=1.15.0",
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "Pillow>=10.0.0",
    "pillow-heif>=0.16.0",
    "numpy>=1.24.0",
]
